refactor(mail): extract shared helpers, add MailStorage class
Address PR #111 review feedback:
- Extract duplicated column/format helpers into command-helpers.ts (DRY)
- Add MailStorage class to encapsulate storage wiring (seen DB + search results)
- Import ALL_COLUMN_KEYS in e2e test instead of hardcoding incomplete list
- Fix JSON extraction in e2e test to handle spinner output correctly
